Kafka on django/djangokafka/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
confluent-kafka==2.0.2
Django==4.1.7
idna==3.4
kafka-python==2.0.2
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
